requirements.txt
torchsummary
torchvision
matplotlib
albumentations
pytorch_lightning
torchmetrics
wandb
einops
timm
gradio
matplotlib
ptwt